+++ apps/hebrew_calendar/app.js
@@ -1,27 +1,1 @@
-g.clear();
-
-let now = new Date();
-let today = require("hebrewDate")(now);
-
-var mainmenu = {
-  "" : {
-    "title" : "Hebrew Date"
-  },
-  cal: {
-    value: require('locale').date(now,1),
-    onchange : () => {}
-  },
-  date: {
-    value : today.date,
-    onchange : () => {}
-  },
-  month: {
-    value : today.month_name,
-    onchange : () => {}
-  },
-  year: {
-    value : today.year,
-    onchange : () => {}
-  }
-};
-E.showMenu(mainmenu);
+!function(){"use strict";g.clear();let e=new Date,a=require("hebrewDate")(e);var n={"":{title:"Hebrew Date"},cal:{value:require("locale").date(e,1),onchange:()=>{}},date:{value:a.date,onchange:()=>{}},month:{value:a.month_name,onchange:()=>{}},year:{value:a.year,onchange:()=>{}}};E.showMenu(n)}();
