+++ apps/hebrew_calendar/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "hebrew_calendar",
+  "version": "0.0.1",
+  "description": "Bangle.js app for seeing hebrew calendar",
+  "main": "app.js",
+  "types": "app.d.ts",
+  "scripts": {
+    "build": "rollup -c"
+  },
+  "author": {
+    "name": "Michael Salaverry",
+    "url": "https://github.com/barakplasma"
+  },
+  "license": "MIT",
+  "devDependencies": {
+    "@rollup/plugin-typescript": "^4.1.1",
+    "rollup": "^2.10.2",
+    "rollup-plugin-terser": "^5.3.0",
+    "terser": "^4.7.0",
+    "tslib": "^2.0.0",
+    "typescript": "^3.9.2"
+  }
+}
