+++ apps/hebrew_calendar/src/app.ts
@@ -1,0 +1,32 @@
+declare var Bangle: any;
+declare var g: any;
+declare var E: any;
+declare var require: any;
+
+g.clear();
+
+let now = new Date();
+let today = require("hebrewDate")(now);
+
+var mainmenu = {
+  "" : {
+    "title" : "Hebrew Date"
+  },
+  cal: {
+    value: require('locale').date(now,1),
+    onchange : () => {}
+  },
+  date: {
+    value : today.date,
+    onchange : () => {}
+  },
+  month: {
+    value : today.month_name,
+    onchange : () => {}
+  },
+  year: {
+    value : today.year,
+    onchange : () => {}
+  }
+};
+E.showMenu(mainmenu);
