+++ apps/hebrew_calendar/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "module": "es2015",
+    "noImplicitAny": true,
+    "target": "es2015"
+  },
+  "include": [
+    "src"
+  ]
+}
